Get historical TRX market data

Retrieves TRX market data for previous days

get_trx_market_data_for_last_n_days(
  vs_currency = "usd",
  days,
  interval = NULL,
  max_attempts = 3L
)

Arguments

vs_currency
(character): name of the base currency to benchmark TRX against (usd by default). An up-to-date list of supported currencies (both fiat and cryptocurrencies) can be retrieved with the get_supported_coingecko_currencies() function. If an unsupported vs_currency is requested, the call will fail with the respective error message.

days
(numeric or "max"): number of days to look back. If the requested number of days covers dates before 2017-11-09, the retrived data will be clipped at 2017-11-09 (the beginning of history for TRX). If days = "max", the entire available history will be retrieved. Depending on the value of days, the time interval used to present the data will differ - see "Details".

interval
(character or NULL): time interval to present the data. The only currently supported interval is daily. Defaults to NULL.

max_attempts
(integer, positive): specifies the maximum number of additional attempts to call a URL if the first attempt fails (i.e. its call status is different from 200). Additional attempts are implemented with an exponential backoff. Defaults to 3.

Value

A tibble with the following columns:

timestamp (POSIXct);

vs_currency (character): same as the argument vs_currency;

price (double): TRX price, as of datetime;

total_trading_vol (double): a 24 h rolling-window trading volume, as of datetime;

market_cap (double): TRX market cap, as of datetime.

Details

If days = 1 and interval = NULL, the data will be presented for every few minutes (typically 3-8 minutes). If days is between 2 and 90 (inclusive) and interval = NULL, an (approximately) hourly time step will be used. Daily data are used for days above 90. If interval = "daily", daily data will be used irrespective of the value of days.

Examples

r <- get_trx_market_data_for_last_n_days(
  vs_currency = "gbp",
  days = 30
)
print(r)

#> # A tibble: 723 x 5
#>    timestamp           vs_currency  price total_trading_vol  market_cap
#>    <dttm>              <chr>        <dbl>             <dbl>       <dbl>
#>  1 2021-01-23 18:44:14 gbp         0.0212        859162835. 1520721687.
#>  2 2021-01-23 19:43:50 gbp         0.0213        856259269. 1529061171.
#>  3 2021-01-23 20:44:29 gbp         0.0212        841149683. 1520877694.
#>  4 2021-01-23 21:43:12 gbp         0.0212        828924507. 1518016367.
#>  5 2021-01-23 22:41:56 gbp         0.0213        823294391. 1522881758.
#>  6 2021-01-23 23:41:14 gbp         0.0213        813531358. 1522240739.
#>  7 2021-01-24 00:49:32 gbp         0.0215        818977798. 1533448680.
#>  8 2021-01-24 01:07:12 gbp         0.0219        849592584. 1555965696.
#>  9 2021-01-24 02:06:11 gbp         0.0216        872872509. 1551294232.
#> 10 2021-01-24 03:06:49 gbp         0.0217        882272016. 1557527044.
#> # … with 713 more rows