Get historical TRX price data

Retrieves TRX open-high-low-close price data for the last n days

get_trx_ohlc_data_for_last_n_days(vs_currency = "usd", days, max_attempts = 3L)

Arguments

vs_currency
(character): name of the base currency to benchmark TRX against (usd by default). An up-to-date list of supported currencies (both fiat and cryptocurrencies) can be retrieved with the get_supported_coingecko_currencies() function. If an unsupported vs_currency is requested, the call will fail with the respective error message.

days
(numeric or "max"): number of days to look back. The only acceptable values are 1, 7, 14, 30, 90, 180, 365 and "max". Attempts to assign any other values will fail with the corresponding console message. If days = "max", the entire available history will be retrieved. If the requested number of days covers dates before 2017-11-09, the retrived data will be clipped at 2017-11-09 (the beginning of history for TRX). Depending on the value of days, the time interval used to present the data will differ - see "Details".

max_attempts
function_params("max_attempts")

Value

A tibble with the following columns:

timestamp (POSIXct): timestamp;

vs_currency (character): same as the argument vs_currency;

price_open (double): TRX price in the beginning of a time iterval;

price_high (double): highest TRX price observed within a time interval;

price_low (double): lowest TRX price observed within a time interval;

price_close (double): TRX price in the end of a time interval.

Details

Granularity of the retrieved data (i.e. candle's body) depends on the number of requested days as follows:

1 day: 30 minutes

7 - 30 days: 4 hours

31 and above: 4 days

Examples

r <- get_trx_ohlc_data_for_last_n_days(days = 7)
print(r)

#> # A tibble: 43 x 6
#>    timestamp           vs_currency price_open price_high price_low price_close
#>    <dttm>              <chr>            <dbl>      <dbl>     <dbl>       <dbl>
#>  1 2021-03-06 20:00:00 usd             0.0505     0.0505    0.0503      0.0503
#>  2 2021-03-07 00:00:00 usd             0.0499     0.0504    0.0497      0.0504
#>  3 2021-03-07 04:00:00 usd             0.0503     0.0507    0.0502      0.0507
#>  4 2021-03-07 08:00:00 usd             0.0505     0.0510    0.0504      0.0507
#>  5 2021-03-07 12:00:00 usd             0.0512     0.0515    0.0510      0.0511
#>  6 2021-03-07 16:00:00 usd             0.0514     0.0515    0.0508      0.0508
#>  7 2021-03-07 20:00:00 usd             0.0510     0.0514    0.0510      0.0512
#>  8 2021-03-08 00:00:00 usd             0.0512     0.0514    0.0510      0.0514
#>  9 2021-03-08 04:00:00 usd             0.0519     0.0519    0.0510      0.0510
#> 10 2021-03-08 08:00:00 usd             0.0510     0.0517    0.0508      0.0508
#> # … with 33 more rows