Get transactions for an account

Returns a list of transactions associated with an account and their attributes

get_tx_info_by_account_address(
  address,
  min_timestamp = 0,
  max_timestamp = NULL,
  add_contract_data = TRUE,
  max_attempts = 3L
)

Arguments

address
(character): address of the account of interest, either in base58check or hex format.

min_timestamp
(numeric or character): a Unix timestamp (including milliseconds), which defines the beginning of the period of interest (inclusive). Defaults to 0.

max_timestamp
(numeric or character): a Unix timestamp (including milliseconds), which defines the end of the period of interest (inclusive).

add_contract_data
(boolean): if TRUE (default), adds column contract_data to the resultant tibble (see Value).

max_attempts
(integer, positive): specifies the maximum number of additional attempts to call a URL if the first attempt fails (i.e. its call status is different from 200). Additional attempts are implemented with an exponential backoff. Defaults to 3.

Value

A nested tibble where each row corresponds to one transaction. A detailed description of the content of this tibble can be found in the help file for get_tx_info_by_id().

Details

As the number of transactions associated with some TRON addresses can be very high, users are advised to choose min_timestamp and max_timestamp wisely. If the query range is too large, the maximum number of transactions returned by the underlying Tronscan API will be capped at 2000. Chunking the time range of interest into smaller periods can help to avoid gaps in data in such cases. However, users would have to implement their own logic for that.

Examples

tx_df <- get_tx_info_by_account_address(
  address = "TAUN6FwrnwwmaEqYcckffC7wYmbaS6cBiX",
  min_timestamp = "1577836800000",
  max_timestamp = "1577838600000"
)

#> 

#> - Fetching data... Elapsed time: 00:00:00

#> 

#> Processing data... [==>-------------------------------------------------]   6%

#> 

#> Processing data... [=====>----------------------------------------------]  11%

#> 

#> Processing data... [========>-------------------------------------------]  17%

#> 

#> Processing data... [===========>----------------------------------------]  22%

#> 

#> Processing data... [=============>--------------------------------------]  28%

#> 

#> Processing data... [================>-----------------------------------]  33%

#> 

#> Processing data... [===================>--------------------------------]  39%

#> 

#> Processing data... [======================>-----------------------------]  44%

#> 

#> Processing data... [=========================>--------------------------]  50%

#> 

#> Processing data... [============================>-----------------------]  56%

#> 

#> Processing data... [===============================>--------------------]  61%

#> 

#> Processing data... [==================================>-----------------]  67%

#> 

#> Processing data... [=====================================>--------------]  72%

#> 

#> Processing data... [=======================================>------------]  78%

#> 

#> Processing data... [==========================================>---------]  83%

#> 

#> Processing data... [=============================================>------]  89%

#> 

#> Processing data... [================================================>---]  94%

#> 

#> Processing data... [====================================================] 100%

#>                                                                               

#> 

print(tx_df)

#> # A tibble: 18 x 21
#>    request_time        address     tx_id        block_number timestamp          
#>    <dttm>              <chr>       <chr>        <chr>        <dttm>             
#>  1 2021-03-24 22:11:41 TAUN6Fwrnw… 36ec1806251… 15860591     2020-01-01 00:00:36
#>  2 2021-03-24 22:11:41 TAUN6Fwrnw… 40fccf97947… 15860591     2020-01-01 00:00:36
#>  3 2021-03-24 22:11:41 TAUN6Fwrnw… c5238b6314d… 15860591     2020-01-01 00:00:36
#>  4 2021-03-24 22:11:40 TAUN6Fwrnw… ce8f3882876… 15860592     2020-01-01 00:00:39
#>  5 2021-03-24 22:11:40 TAUN6Fwrnw… eff1e3312c3… 15860693     2020-01-01 00:05:42
#>  6 2021-03-24 22:11:40 TAUN6Fwrnw… df70e530c1a… 15860693     2020-01-01 00:05:42
#>  7 2021-03-24 22:11:40 TAUN6Fwrnw… 232c646de3f… 15860693     2020-01-01 00:05:42
#>  8 2021-03-24 22:11:39 TAUN6Fwrnw… 9a167921db7… 15860694     2020-01-01 00:05:45
#>  9 2021-03-24 22:11:38 TAUN6Fwrnw… 25e74261b1e… 15860816     2020-01-01 00:11:51
#> 10 2021-03-24 22:11:39 TAUN6Fwrnw… 30035b98f4a… 15860816     2020-01-01 00:11:51
#> 11 2021-03-24 22:11:39 TAUN6Fwrnw… 43d7d11a21f… 15860816     2020-01-01 00:11:51
#> 12 2021-03-24 22:11:39 TAUN6Fwrnw… 11bd89eafe0… 15860816     2020-01-01 00:11:51
#> 13 2021-03-24 22:11:38 TAUN6Fwrnw… 6f31221759e… 15860895     2020-01-01 00:15:48
#> 14 2021-03-24 22:11:37 TAUN6Fwrnw… 6c95eaf190c… 15861004     2020-01-01 00:21:15
#> 15 2021-03-24 22:11:38 TAUN6Fwrnw… 106d707772b… 15861004     2020-01-01 00:21:15
#> 16 2021-03-24 22:11:37 TAUN6Fwrnw… 514d501b733… 15861115     2020-01-01 00:26:48
#> 17 2021-03-24 22:11:36 TAUN6Fwrnw… 2745b8012f0… 15861177     2020-01-01 00:29:54
#> 18 2021-03-24 22:11:37 TAUN6Fwrnw… aae2d97414d… 15861177     2020-01-01 00:29:54
#> # … with 16 more variables: contract_result <chr>, confirmed <lgl>,
#> #   confirmations_count <int>, sr_confirm_list <list>, contract_type <chr>,
#> #   from_address <chr>, to_address <chr>, is_contract_from_address <lgl>,
#> #   is_contract_to_address <lgl>, costs <list>, trx_transfer <dbl>,
#> #   trc10_transfer <list>, trc20_transfer <lgl>, internal_tx <lgl>, info <lgl>,
#> #   contract_data <list>