List top TRX holders

Returns a list of accounts with the largest TRX balances

list_top_trx_holders(n = 20, max_attempts = 3L)

Arguments

n
(double): number of top accounts to retrieve.

max_attempts
(integer, positive): specifies the maximum number of additional attempts to call a URL if the first attempt fails (i.e. its call status is different from 200). Additional attempts are implemented with an exponential backoff. Defaults to 3.

Value

A tibble with n rows and the following columns:

request_time (POSIXct): date and time when the request was made;

address (character): account address (in base58check format);

trx_balance (double): TRX balance;

total_tx (integer): total number of transactions associated with the respective account;

tron_power (double): amount of TRX frozen (see official documentation for details).

Examples

r <- list_top_trx_holders(10)
print(r)

#> # A tibble: 10 x 5
#>    request_time        address                   trx_balance total_tx tron_power
#>    <dttm>              <chr>                           <dbl>    <int>      <dbl>
#>  1 2021-07-12 22:04:28 TNUC9Qb1rRpS5CbWLmNMxXBj…     1.66e10  1701130          0
#>  2 2021-07-12 22:04:28 TA9FnQrLGdgLW6cwBKue9Dyq…     2.63e 9   126582   42000000
#>  3 2021-07-12 22:04:28 TE2RzoSV3wFK99w6J9UnnZ4v…     1.96e 9   378918          0
#>  4 2021-07-12 22:04:28 TPyjyZfsYaXStgz2NmAraF1u…     1.68e 9      270          0
#>  5 2021-07-12 22:04:28 TNaRAoLUyYEV2uF7GUrzSjRQ…     1.57e 9 25916908    6002000
#>  6 2021-07-12 22:04:28 TTd9qHyjqiUkfTxe3gotbuTM…     1.32e 9   401466          0
#>  7 2021-07-12 22:04:28 TQn9Y2khEsLJW1ChVWFMSMeR…     1.29e 9  5992614          0
#>  8 2021-07-12 22:04:28 TMuA6YqfCeX8EhbfYEg5y7S4…     1.06e 9     1578 6442761345
#>  9 2021-07-12 22:04:28 TUTtLtGjXWqkTAow7jj7oMhT…     9.09e 8      698          0
#> 10 2021-07-12 22:04:28 TH2mEwTKNgtg8psR6Qx2RBUX…     8.68e 8   704965          0