Parse event attributes

Converts a list with event attributes into a nested tibble

parse_events_info(info)

Arguments

info
A non-empty, named list returned as a result of calling events-related methods (see get_events_by_block_number(), get_events_by_contract_address(), etc.).

Value

A nested tibble where each row corresponds to an event associated with the account of interest. This tibble contains the following columns:

tx_id (character): transaction ID;

block_number (character);

timestamp (POSIXct, UTC timezone): block timestamp;

contract_address (character): adress of the smart contract that triggered the event;

event_name (character): possible values of this column are contract- and event-specific;

event_data (list): each element of this list contains a named list with additional attributes of the event.