Validate arguments

Validates arguments that are commonly used in tronr functions

validate_arguments(
  arg_address = NULL,
  arg_contract_address = NULL,
  arg_related_address = NULL,
  arg_token_id = NULL,
  arg_token_name = NULL,
  arg_tx_id = NULL,
  arg_event_name = NULL,
  arg_block_number = NULL,
  arg_only_confirmed = NULL,
  arg_only_unconfirmed = NULL,
  arg_only_to = NULL,
  arg_only_from = NULL,
  arg_min_timestamp = NULL,
  arg_max_timestamp = NULL,
  arg_detailed_info = NULL,
  arg_add_contract_data = NULL,
  arg_order_by = NULL,
  arg_direction = NULL,
  arg_max_attempts = NULL,
  arg_vs_currencies = NULL
)

Arguments

arg_address
(character): account or smart contract address;

arg_contract_address
(character): contract address;

arg_related_address
(character): account of interest in the context of a given query;

arg_token_id
(character): token ID;

arg_token_name
(character): token name;

arg_tx_id
(character): transaction ID;

arg_event_name
(character): event name;

arg_block_number
(character): block number;

arg_only_confirmed
(boolean): whether results should be returned for confirmed transactions only;

arg_only_unconfirmed
(boolean): whether results should be returned for unconfirmed transactions only;

arg_only_to
(boolean): whether results should be returned for transactions directed to a given account;

arg_only_from
(boolean): whether results should be returned for transactions originated from a given account;

arg_min_timestamp
(numeric or character): start of the time range to retrieve the data from;

arg_max_timestamp
(numeric or character): end of the time range to retrieve the data from;

arg_detailed_info
(boolean): whether detailed information should be retrieved for an entity (e.g. TRC-10 tokens);

arg_add_contract_data
(boolean): whether to add raw contract data to the tibble with transaction attributes.

arg_order_by
(character): variable to order the results by;

arg_direction
(character): how to order the results (desc or asc);

arg_max_attempts
(integer): number of additional attempts to call the API if the first attempt fails.

arg_vs_currencies
(character): abbreviated names of the currencies to benchmark TRX against (using the public CoinGecko API).

Value

This function is to be used for its side effects. If any of the checks is not passed, it returns the respective error. If all checks are passed successfully, nothing is returned.