Get historical TRX market data

Retrieves TRX market data for previous days

get_trx_market_data_for_last_n_days(
  vs_currency = "usd",
  days,
  interval = NULL,
  max_attempts = 3L
)

Arguments

vs_currency
(character): name of the base currency to benchmark TRX against (usd by default). An up-to-date list of supported currencies (both fiat and cryptocurrencies) can be retrieved with the get_supported_coingecko_currencies() function. If an unsupported vs_currency is requested, the call will fail with the respective error message.

days
(numeric or "max"): number of days to look back. If the requested number of days covers dates before 2017-11-09, the retrived data will be clipped at 2017-11-09 (the beginning of history for TRX). If days = "max", the entire available history will be retrieved. Depending on the value of days, the time interval used to present the data will differ - see "Details".

interval
(character or NULL): time interval to present the data. The only currently supported interval is daily. Defaults to NULL.

max_attempts
(integer, positive): specifies the maximum number of additional attempts to call a URL if the first attempt fails (i.e. its call status is different from 200). Additional attempts are implemented with an exponential backoff. Defaults to 3.

Value

A tibble with the following columns:

timestamp (POSIXct);

vs_currency (character): same as the argument vs_currency;

price (double): TRX price, as of datetime;

total_trading_vol (double): a 24 h rolling-window trading volume, as of datetime;

market_cap (double): TRX market cap, as of datetime.

Details

If days = 1 and interval = NULL, the data will be presented for every few minutes (typically 3-8 minutes). If days is between 2 and 90 (inclusive) and interval = NULL, an (approximately) hourly time step will be used. Daily data are used for days above 90. If interval = "daily", daily data will be used irrespective of the value of days.

Examples

r <- get_trx_market_data_for_last_n_days(
  vs_currency = "gbp",
  days = 30
)
print(r)

#> # A tibble: 721 x 5
#>    timestamp           vs_currency  price total_trading_vol  market_cap
#>    <dttm>              <chr>        <dbl>             <dbl>       <dbl>
#>  1 2021-02-22 22:04:46 gbp         0.0393       2135941289. 2810628844.
#>  2 2021-02-22 23:07:43 gbp         0.0386       2099911426. 2764768976.
#>  3 2021-02-23 00:05:10 gbp         0.0388       2124964811. 2776822517.
#>  4 2021-02-23 01:03:43 gbp         0.0382       2103649745. 2734607069.
#>  5 2021-02-23 02:04:32 gbp         0.0368       2057690047. 2665418532.
#>  6 2021-02-23 03:06:45 gbp         0.0358       2045291572. 2563543646.
#>  7 2021-02-23 04:09:45 gbp         0.0367       2124656402. 2628498951.
#>  8 2021-02-23 05:09:16 gbp         0.0355       2058691663. 2542263094.
#>  9 2021-02-23 06:05:25 gbp         0.0352       2107069274. 2524884667.
#> 10 2021-02-23 07:08:11 gbp         0.0346       2086663635. 2479476130.
#> # … with 711 more rows