Get historical TRX market data

Retrieves TRX market data for previous days

get_trx_market_data_for_last_n_days(
  vs_currency = "usd",
  days,
  interval = NULL,
  max_attempts = 3L
)

Arguments

vs_currency
(character): name of the base currency to benchmark TRX against (usd by default). An up-to-date list of supported currencies (both fiat and cryptocurrencies) can be retrieved with the get_supported_coingecko_currencies() function. If an unsupported vs_currency is requested, the call will fail with the respective error message.

days
(numeric or "max"): number of days to look back. If the requested number of days covers dates before 2017-11-09, the retrived data will be clipped at 2017-11-09 (the beginning of history for TRX). If days = "max", the entire available history will be retrieved. Depending on the value of days, the time interval used to present the data will differ - see "Details".

interval
(character or NULL): time interval to present the data. The only currently supported interval is daily. Defaults to NULL.

max_attempts
(integer, positive): specifies the maximum number of additional attempts to call a URL if the first attempt fails (i.e. its call status is different from 200). Additional attempts are implemented with an exponential backoff. Defaults to 3.

Value

A tibble with the following columns:

timestamp (POSIXct);

vs_currency (character): same as the argument vs_currency;

price (double): TRX price, as of datetime;

total_trading_vol (double): a 24 h rolling-window trading volume, as of datetime;

market_cap (double): TRX market cap, as of datetime.

Details

If days = 1 and interval = NULL, the data will be presented for every few minutes (typically 3-8 minutes). If days is between 2 and 90 (inclusive) and interval = NULL, an (approximately) hourly time step will be used. Daily data are used for days above 90. If interval = "daily", daily data will be used irrespective of the value of days.

Examples

r <- get_trx_market_data_for_last_n_days(
  vs_currency = "gbp",
  days = 30
)
print(r)

#> # A tibble: 722 x 5
#>    timestamp           vs_currency  price total_trading_vol  market_cap
#>    <dttm>              <chr>        <dbl>             <dbl>       <dbl>
#>  1 2021-01-22 15:19:36 gbp         0.0211       1294483634. 1505710308.
#>  2 2021-01-22 16:13:35 gbp         0.0210       1276713326. 1504099608.
#>  3 2021-01-22 17:00:36 gbp         0.0209       1267602035. 1502204704.
#>  4 2021-01-22 18:01:30 gbp         0.0208       1254982344. 1487155336.
#>  5 2021-01-22 19:01:14 gbp         0.0212       1269456897. 1511242690.
#>  6 2021-01-22 20:04:46 gbp         0.0210       1258936472. 1506762444.
#>  7 2021-01-22 21:20:22 gbp         0.0210       1243496608. 1502792440.
#>  8 2021-01-22 22:01:46 gbp         0.0210       1234827632. 1504459248.
#>  9 2021-01-22 23:24:47 gbp         0.0209       1213951158. 1495086079.
#> 10 2021-01-23 00:27:47 gbp         0.0207       1196695738. 1485689056.
#> # … with 712 more rows