Get historical TRX market data

Retrieves TRX market data for previous days

get_trx_market_data_for_last_n_days(
  vs_currency = "usd",
  days,
  interval = NULL,
  max_attempts = 3L
)

Arguments

vs_currency
(character): name of the base currency to benchmark TRX against (usd by default). An up-to-date list of supported currencies (both fiat and cryptocurrencies) can be retrieved with the get_supported_coingecko_currencies() function. If an unsupported vs_currency is requested, the call will fail with the respective error message.

days
(numeric or "max"): number of days to look back. If the requested number of days covers dates before 2017-11-09, the retrived data will be clipped at 2017-11-09 (the beginning of history for TRX). If days = "max", the entire available history will be retrieved. Depending on the value of days, the time interval used to present the data will differ - see "Details".

interval
(character or NULL): time interval to present the data. The only currently supported interval is daily. Defaults to NULL.

max_attempts
(integer, positive): specifies the maximum number of additional attempts to call a URL if the first attempt fails (i.e. its call status is different from 200). Additional attempts are implemented with an exponential backoff. Defaults to 3.

Value

A tibble with the following columns:

timestamp (POSIXct);

vs_currency (character): same as the argument vs_currency;

price (double): TRX price, as of datetime;

total_trading_vol (double): a 24 h rolling-window trading volume, as of datetime;

market_cap (double): TRX market cap, as of datetime.

Details

If days = 1 and interval = NULL, the data will be presented for every few minutes (typically 3-8 minutes). If days is between 2 and 90 (inclusive) and interval = NULL, an (approximately) hourly time step will be used. Daily data are used for days above 90. If interval = "daily", daily data will be used irrespective of the value of days.

Examples

r <- get_trx_market_data_for_last_n_days(
  vs_currency = "gbp",
  days = 30
)
print(r)

#> # A tibble: 721 x 5
#>    timestamp           vs_currency  price total_trading_vol  market_cap
#>    <dttm>              <chr>        <dbl>             <dbl>       <dbl>
#>  1 2021-02-11 18:07:18 gbp         0.0354       1780074776. 2537800058.
#>  2 2021-02-11 19:02:12 gbp         0.0353       1808876926. 2540166762.
#>  3 2021-02-11 20:01:34 gbp         0.0367       1966969388. 2603753623.
#>  4 2021-02-11 21:12:48 gbp         0.0377       2178586176. 2657913768.
#>  5 2021-02-11 22:07:09 gbp         0.0373       2233572720. 2672515691.
#>  6 2021-02-11 23:09:47 gbp         0.0394       2398958576. 2823653076.
#>  7 2021-02-12 00:07:32 gbp         0.0412       2670598788. 2951986898.
#>  8 2021-02-12 01:24:29 gbp         0.0401       2642747418. 2887573532.
#>  9 2021-02-12 02:07:22 gbp         0.0402       2636366215. 2891706310.
#> 10 2021-02-12 03:07:21 gbp         0.0395       2664050828. 2828904896.
#> # … with 711 more rows