List top TRX holders

Returns a list of accounts with the largest TRX balances

list_top_trx_holders(n = 20, max_attempts = 3L)

Arguments

n
(double): number of top accounts to retrieve.

max_attempts
(integer, positive): specifies the maximum number of additional attempts to call a URL if the first attempt fails (i.e. its call status is different from 200). Additional attempts are implemented with an exponential backoff. Defaults to 3.

Value

A tibble with n rows and the following columns:

request_time (POSIXct): date and time when the request was made;

address (character): account address (in base58check format);

trx_balance (double): TRX balance;

total_tx (integer): total number of transactions associated with the respective account;

tron_power (double): amount of TRX frozen (see official documentation for details).

Examples

r <- list_top_trx_holders(10)
print(r)

#> # A tibble: 10 x 5
#>    request_time        address                   trx_balance total_tx tron_power
#>    <dttm>              <chr>                           <dbl>    <int>      <dbl>
#>  1 2021-03-13 18:01:36 TNUC9Qb1rRpS5CbWLmNMxXBj…     1.52e10  1360892          0
#>  2 2021-03-13 18:01:36 TQcia2H2TU3WrFk9sKtdK9qC…     2.86e 9   286279          0
#>  3 2021-03-13 18:01:36 TE2RzoSV3wFK99w6J9UnnZ4v…     2.43e 9   149727          0
#>  4 2021-03-13 18:01:36 TA9FnQrLGdgLW6cwBKue9Dyq…     2.42e 9    59338   23000000
#>  5 2021-03-13 18:01:36 TNaRAoLUyYEV2uF7GUrzSjRQ…     1.74e 9 13457808          0
#>  6 2021-03-13 18:01:36 TQn9Y2khEsLJW1ChVWFMSMeR…     1.39e 9  3242116          0
#>  7 2021-03-13 18:01:36 TWd4WrZ9wn84f5x1hZhL4DHv…     1.20e 9      604 3239581502
#>  8 2021-03-13 18:01:36 TTd9qHyjqiUkfTxe3gotbuTM…     1.18e 9    88819          0
#>  9 2021-03-13 18:01:36 TKAtLoCB529zusLfLVkGvLNi…     1.06e 9   215817          0
#> 10 2021-03-13 18:01:36 TH2mEwTKNgtg8psR6Qx2RBUX…     7.40e 8   256552          0